Move deprecated issue-related functions to base class
