+++ ogr/services/gitlab/project.py
@@ -30,7 +30,6 @@ from ogr.abstract import (
     PullRequest,
     Issue,
     Release,
-    IssueComment,
     PRComment,
     GitTag,
     IssueStatus,
@@ -42,7 +41,7 @@ from ogr.exceptions import GitlabAPIException
 from ogr.services import gitlab as ogr_gitlab
 from ogr.services.base import BaseGitProject
 from ogr.services.gitlab.release import GitlabRelease
-from ogr.services.gitlab.comments import GitlabIssueComment, GitlabPRComment
+from ogr.services.gitlab.comments import GitlabPRComment
 from ogr.services.gitlab.issue import GitlabIssue
 
 logger = logging.getLogger(__name__)
@@ -176,10 +175,6 @@ class GitlabProject(BaseGitProject):
             )
         )
 
-    def can_close_issue(self, username: str, issue: Issue) -> bool:
-        # TODO: deprecate
-        return issue.can_close_issue(username)
-
     def can_merge_pr(self, username) -> bool:
         allowed_users = self.who_can_close_issue()
 
@@ -206,27 +201,6 @@ class GitlabProject(BaseGitProject):
             for member in self.gitlab_repo.members.all(all=True)
             if member.access_level in access_levels
         ]
-
-    def _get_all_issue_comments(self, issue_id) -> List["IssueComment"]:
-        # TODO: deprecate
-        issue = self.gitlab_repo.issues.get(issue_id)
-
-        return [
-            GitlabIssueComment(raw_comment)
-            for raw_comment in issue.notes.list(sort="asc")
-        ]
-
-    def issue_close(self, issue_id: int) -> Issue:
-        # TODO: deprecate
-        return self.get_issue(issue_id).close()
-
-    def get_issue_labels(self, issue_id: int) -> List[str]:
-        # TODO: deprecate
-        return self.get_issue(issue_id).get_labels()
-
-    def add_issue_labels(self, issue_id, labels) -> None:
-        # TODO: deprecate
-        self.get_issue(issue_id).add_labels(labels)
 
     def get_pr_list(self, status: PRStatus = PRStatus.open) -> List["PullRequest"]:
         # Gitlab API has status 'opened', not 'open'
@@ -425,22 +399,11 @@ class GitlabProject(BaseGitProject):
             logger.error(f"Issue {issue_id} was not found.")
             raise GitlabAPIException(f"Issue {issue_id} was not found. ", ex)
 
-    def get_issue_info(self, issue_id: int) -> Issue:
-        # TODO: deprecate
-        return self.get_issue(issue_id)
-
     def create_issue(self, title: str, description: str) -> Issue:
         issue = self.gitlab_repo.issues.create(
             {"title": title, "description": description}
         )
         return GitlabIssue(issue, self)
-
-    def issue_comment(self, issue_id: int, body: str) -> IssueComment:
-        """
-        Create comment on an issue.
-        """
-        # TODO: deprecate
-        return self.get_issue(issue_id).issue_comment(body)
 
     def get_pr_info(self, pr_id: int) -> PullRequest:
         mr = self.gitlab_repo.mergerequests.get(pr_id)
